TodoEndpoint some many changes
